Move IPolicyRequirementFactory to separate file
